Max/Min Heap refactor to a abstract parent class
Changed heapProofs to BasicMultipleChoiceQuestion.ts
Added a second proof
