+++ settings/questionsSelection.ts
@@ -2,7 +2,11 @@
  * This file contains a list of all questions available on the give instance of SEAL.
  * TODO: In future this file should be generated by a script.
  */
-import { HeapOperations, HeapUnderstanding } from "@shared/question-generators/heap/heapGenerator.ts"
+import {
+  HeapOperations,
+  HeapProofs,
+  HeapUnderstanding,
+} from "@shared/question-generators/heap/heapGenerator.ts"
 import { huffmanCoding } from "@shared/question-generators/huffman-coding/huffmanCoding.ts"
 import { QuestionRoutes } from "../shared/src/api/QuestionRouter"
 // import { ExampleQuestion } from "../shared/src/question-generators/example/example"
@@ -58,12 +62,16 @@ export const allQuestionGeneratorRoutes: QuestionRoutes = [
     generator: huffmanCoding,
   },
   {
-    path: "heap/heap",
+    path: "heap/heapOperation",
     generator: HeapOperations,
   },
   {
     path: "heap/heapUnderstanding",
     generator: HeapUnderstanding,
+  },
+  {
+    path: "heap/heapProof",
+    generator: HeapProofs,
   },
 ]
